refactor(provider): use ontology OCR types directly, move check_response to HttpClient
- OCR backends import Word/Line/Block/Page directly from
  nvisy_ontology::artifacts instead of re-exporting through the
  provider's output module
- Remove duplicate re-export chains for OCR structural types
- Move check_response from ocr::backend to HttpClient::check_response
