+++ crates/nvisy-ontology/src/artifacts/image.rs
@@ -40,7 +40,17 @@ pub struct Line {
 }
 
 /// Classification of a block region.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    JsonSchema
+)]
 #[serde(rename_all = "snake_case")]
 pub enum BlockKind {
     /// Paragraph / prose.
