+++ crates/nvisy-provider/src/agent/mod.rs
@@ -22,7 +22,7 @@ pub use self::ner::{
     KnownNerEntity, NerAgent, NerContext, NerEntities, NerEntity, ResolvedOffsets,
 };
 pub use self::ocr::{
-    Backend, Block, BlockKind, ImageFormat, Line, OcrAgent, OcrProvider, Page, ProposedEntity,
-    VerificationCandidate, VerificationOutput, VerificationStatus, VerifiedEntity, Word,
+    Backend, ImageFormat, OcrAgent, OcrProvider, ProposedEntity, VerificationCandidate,
+    VerificationOutput, VerificationStatus, VerifiedEntity,
 };
 pub use crate::ocr::{ImageInput, ImageOutput, RunParams};
