+++ crates/nvisy-provider/src/ocr/provider/aws_textract/backend.rs
@@ -7,15 +7,14 @@ use std::fmt;
 
 use hmac::{Hmac, KeyInit, Mac};
 use nvisy_core::{Error, Result};
+use nvisy_ontology::artifacts::{Block, BlockKind, Line, Page, Word};
 use nvisy_ontology::math::{BoundingBox, Polygon, Vertex};
 use serde::Deserialize;
 use sha2::{Digest, Sha256};
 
 use super::AwsTextractParams;
 use crate::http::{HttpClient, HttpConfig};
-use crate::ocr::backend::{
-    Backend, Block, BlockKind, ImageInput, ImageOutput, Line, Page, RunParams, Word, check_response,
-};
+use crate::ocr::backend::{Backend, ImageInput, ImageOutput, RunParams};
 
 /// [`Backend`] implementation for AWS Textract.
 ///
@@ -250,7 +249,7 @@ impl Backend for AwsTextractBackend {
             .await
             .map_err(|e| Error::connection(e.to_string(), "aws_textract_ocr", true))?;
 
-        let resp = check_response(resp, "Textract").await?;
+        let resp = HttpClient::check_response(resp, "Textract").await?;
 
         let parsed: TextractResponse = resp.json().await.map_err(|e| {
             Error::runtime(
