+++ crates/nvisy-provider/src/ocr/provider/datalab_surya/backend.rs
@@ -3,16 +3,14 @@
 //! [`Backend`]: crate::Backend
 
 use nvisy_core::{Error, Result};
+use nvisy_ontology::artifacts::{Block, BlockKind, Line, Page, Word};
 use nvisy_ontology::math::{BoundingBox, Polygon, Vertex};
 use reqwest_middleware::reqwest::multipart::Form;
 use serde::Deserialize;
 
 use super::SuryaParams;
 use crate::http::{HttpClient, HttpConfig};
-use crate::ocr::backend::{
-    Backend, Block, BlockKind, ImageInput, ImageOutput, Line, Page, RunParams, Word,
-    check_response, image_part,
-};
+use crate::ocr::backend::{Backend, ImageInput, ImageOutput, RunParams, image_part};
 
 /// [`Backend`] implementation for Surya OCR.
 ///
@@ -96,7 +94,7 @@ impl Backend for SuryaBackend {
             .await
             .map_err(|e| Error::connection(e.to_string(), "surya_ocr", true))?;
 
-        let resp = check_response(resp, "Surya").await?;
+        let resp = HttpClient::check_response(resp, "Surya").await?;
 
         let parsed: SuryaResponse = resp.json().await.map_err(|e| {
             Error::runtime(format!("Surya JSON parse error: {e}"), "surya_ocr", false)
