+++ crates/nvisy-provider/src/ocr/provider/google_vision/backend.rs
@@ -5,14 +5,13 @@
 use std::fmt;
 
 use nvisy_core::{Error, Result};
+use nvisy_ontology::artifacts::{Block, BlockKind, Line, Page, Word};
 use nvisy_ontology::math::{BoundingBox, Polygon, Vertex};
 use serde::Deserialize;
 
 use super::GoogleVisionParams;
 use crate::http::{HttpClient, HttpConfig};
-use crate::ocr::backend::{
-    Backend, Block, BlockKind, ImageInput, ImageOutput, Line, Page, RunParams, Word, check_response,
-};
+use crate::ocr::backend::{Backend, ImageInput, ImageOutput, RunParams};
 
 /// [`Backend`] implementation for Google Cloud Vision API.
 ///
@@ -158,7 +157,7 @@ impl Backend for GoogleVisionBackend {
             .await
             .map_err(|e| Error::connection(e.to_string(), "google_vision_ocr", true))?;
 
-        let resp = check_response(resp, "Google Vision").await?;
+        let resp = HttpClient::check_response(resp, "Google Vision").await?;
 
         let parsed: AnnotateResponse = resp.json().await.map_err(|e| {
             Error::runtime(
